+++ x/btcstaking/keeper/btc_staking_consumer_events_test.go
@@ -124,7 +124,9 @@ func FuzzSetBTCStakingEventStore_ActiveDel(f *testing.F) {
 		// delegation related assertions
 		actualDel, err := h.BTCStakingKeeper.GetBTCDelegation(h.Ctx, stakingTxHash)
 		h.NoError(err)
-		require.False(t, actualDel.HasCovenantQuorums(h.BTCStakingKeeper.GetParams(h.Ctx).CovenantQuorum))
+		hasQuorum, err := h.BTCStakingKeeper.BtcDelHasCovenantQuorums(h.Ctx, actualDel, h.BTCStakingKeeper.GetParams(h.Ctx).CovenantQuorum)
+		h.NoError(err)
+		require.False(t, hasQuorum)
 		// create cov sigs to activate the delegation
 		msgs := h.GenerateCovenantSignaturesMessages(r, covenantSKs, msgCreateBTCDel, actualDel)
 		bogusMsg := *msgs[0]
@@ -138,11 +140,11 @@ func FuzzSetBTCStakingEventStore_ActiveDel(f *testing.F) {
 		// ensure the BTC delegation now has voting power as it has been activated
 		actualDel, err = h.BTCStakingKeeper.GetBTCDelegation(h.Ctx, stakingTxHash)
 		h.NoError(err)
-		require.True(t, actualDel.HasCovenantQuorums(h.BTCStakingKeeper.GetParams(h.Ctx).CovenantQuorum))
+		require.True(t, actualDel.HasCovenantQuorums(h.BTCStakingKeeper.GetParams(h.Ctx).CovenantQuorum, 0))
 		require.True(t, actualDel.BtcUndelegation.HasCovenantQuorums(h.BTCStakingKeeper.GetParams(h.Ctx).CovenantQuorum))
 		votingPower := actualDel.VotingPower(
 			h.BTCLightClientKeeper.GetTipInfo(h.Ctx).Height,
-			h.BTCStakingKeeper.GetParams(h.Ctx).CovenantQuorum,
+			h.BTCStakingKeeper.GetParams(h.Ctx).CovenantQuorum, 0,
 		)
 		require.Equal(t, uint64(stakingValue), votingPower)
 
@@ -225,7 +227,8 @@ func FuzzSetBTCStakingEventStore_UnbondedDel(f *testing.F) {
 		actualDel, err = h.BTCStakingKeeper.GetBTCDelegation(h.Ctx, stakingTxHash)
 		h.NoError(err)
 		btcTip := uint32(30)
-		status := actualDel.GetStatus(btcTip, bsParams.CovenantQuorum)
+		status, err := h.BTCStakingKeeper.BtcDelStatus(h.Ctx, actualDel, bsParams.CovenantQuorum, btcTip)
+		h.NoError(err)
 		require.Equal(t, types.BTCDelegationStatus_ACTIVE, status)
 
 		msg := &types.MsgBTCUndelegate{
@@ -276,7 +279,9 @@ func FuzzSetBTCStakingEventStore_UnbondedDel(f *testing.F) {
 		// ensure the BTC delegation is unbonded
 		actualDel, err = h.BTCStakingKeeper.GetBTCDelegation(h.Ctx, stakingTxHash)
 		h.NoError(err)
-		status = actualDel.GetStatus(btcTip, bsParams.CovenantQuorum)
+
+		status, err = h.BTCStakingKeeper.BtcDelStatus(h.Ctx, actualDel, bsParams.CovenantQuorum, btcTip)
+		h.NoError(err)
 		require.Equal(t, types.BTCDelegationStatus_UNBONDED, status)
 
 		// event store related assertions
@@ -389,6 +394,20 @@ func TestDeterministicOrdering(t *testing.T) {
 	h.GenAndApplyParams(r)
 
 	bsnIds := []string{"bsn-z", "bsn-a", "bsn-m", "bsn-b"}
+
+	// register consumers as cosmos consumers
+	for _, bsnId := range bsnIds {
+		consumerRegister := &bsctypes.ConsumerRegister{
+			ConsumerId: bsnId,
+			ConsumerMetadata: &bsctypes.ConsumerRegister_CosmosConsumerMetadata{
+				CosmosConsumerMetadata: &bsctypes.CosmosConsumerMetadata{
+					ChannelId: bsnId,
+				},
+			},
+		}
+		err := h.BTCStkConsumerKeeper.RegisterConsumer(h.Ctx, consumerRegister)
+		require.NoError(t, err)
+	}
 
 	// Add consumer events in random order
 	for _, bsnId := range bsnIds {