+++ x/btcstaking/types/errors.go
@@ -35,4 +35,6 @@ var (
 	ErrNoBabylonFPRestaked       = errorsmod.Register(ModuleName, 1126, "BTC delegation must include at least one Babylon Genesis finality provider in its multi-staking selection")
 	ErrEmptyCommissionRates      = errorsmod.Register(ModuleName, 1127, "empty commission")
 	ErrLargestBtcReorgNotFound   = errorsmod.Register(ModuleName, 1128, "there is no BTC reorg currently set")
+	ErrFpBSNIdNotRegistered      = errorsmod.Register(ModuleName, 1129, "the BSN id selected is not registered")
+	ErrInvalidStakeExpansion     = errorsmod.Register(ModuleName, 1130, "invalid stake expansion")
 )