+++ x/btcstkconsumer/client/cli/query.go
@@ -25,9 +25,6 @@ func GetQueryCmd(queryRoute string) *cobra.Command {
 	cmd.AddCommand(CmdQueryParams())
 	cmd.AddCommand(CmdRegisteredConsumer())
 	cmd.AddCommand(CmdRegisteredConsumers())
-	cmd.AddCommand(CmdFinalityProviderConsumer())
-	cmd.AddCommand(CmdFinalityProvider())
-	cmd.AddCommand(CmdFinalityProviders())
 
 	return cmd
 }
@@ -87,94 +84,6 @@ func CmdRegisteredConsumer() *cobra.Command {
 	}
 
 	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
-
-func CmdFinalityProviderConsumer() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "finality-provider-consumer <fp_btc_pk_hex>",
-		Short: "retrieve a given Consumer finality provider's registered consumer id",
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx := client.GetClientContextFromCmd(cmd)
-			queryClient := types.NewQueryClient(clientCtx)
-			res, err := queryClient.FinalityProviderConsumer(
-				cmd.Context(),
-				&types.QueryFinalityProviderConsumerRequest{
-					FpBtcPkHex: args[0],
-				},
-			)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
-
-func CmdFinalityProvider() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "finality-provider <consumer-id> <fp_btc_pk_hex>",
-		Short: "retrieve a given consumer's finality provider",
-		Args:  cobra.ExactArgs(2),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx := client.GetClientContextFromCmd(cmd)
-			queryClient := types.NewQueryClient(clientCtx)
-			res, err := queryClient.FinalityProvider(
-				cmd.Context(),
-				&types.QueryFinalityProviderRequest{
-					FpBtcPkHex: args[1],
-					ConsumerId: args[0],
-				},
-			)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
-
-func CmdFinalityProviders() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "finality-providers <consumer-id>",
-		Short: "retrieve a given consumer's all finality providers",
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx := client.GetClientContextFromCmd(cmd)
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			pageReq, err := client.ReadPageRequest(cmd.Flags())
-			if err != nil {
-				return err
-			}
-
-			res, err := queryClient.FinalityProviders(cmd.Context(), &types.QueryFinalityProvidersRequest{
-				ConsumerId: args[0],
-				Pagination: pageReq,
-			})
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-	flags.AddPaginationFlagsToCmd(cmd, "finality-providers")
 
 	return cmd
 }