+++ x/btcstkconsumer/keeper/grpc_query.go
@@ -3,10 +3,6 @@ package keeper
 import (
 	"context"
 
-	errorsmod "cosmossdk.io/errors"
-	btcstaking "github.com/babylonlabs-io/babylon/v3/x/btcstaking/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-
 	bbn "github.com/babylonlabs-io/babylon/v3/types"
 	"github.com/babylonlabs-io/babylon/v3/x/btcstkconsumer/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -82,82 +78,4 @@ func (k Keeper) ConsumersRegistry(c context.Context, req *types.QueryConsumersRe
 
 	resp := &types.QueryConsumersRegistryResponse{ConsumerRegisters: consumersRegisters}
 	return resp, nil
-}
-
-// FinalityProviders returns a paginated list of all registered finality providers for a given consumer
-func (k Keeper) FinalityProviders(c context.Context, req *types.QueryFinalityProvidersRequest) (*types.QueryFinalityProvidersResponse, error) {
-	if req == nil {
-		return nil, status.Error(codes.InvalidArgument, "empty request")
-	}
-
-	ctx := sdk.UnwrapSDKContext(c)
-	store := k.finalityProviderStore(ctx, req.ConsumerId)
-
-	var fpResp []*types.FinalityProviderResponse
-	pageRes, err := query.Paginate(store, req.Pagination, func(key, value []byte) error {
-		var fp btcstaking.FinalityProvider
-		if err := fp.Unmarshal(value); err != nil {
-			return err
-		}
-
-		resp := types.NewFinalityProviderResponse(&fp)
-		fpResp = append(fpResp, resp)
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return &types.QueryFinalityProvidersResponse{FinalityProviders: fpResp, Pagination: pageRes}, nil
-}
-
-// FinalityProvider returns the finality provider with the specified finality provider BTC PK
-func (k Keeper) FinalityProvider(c context.Context, req *types.QueryFinalityProviderRequest) (*types.QueryFinalityProviderResponse, error) {
-	if req == nil {
-		return nil, status.Error(codes.InvalidArgument, "empty request")
-	}
-
-	if len(req.FpBtcPkHex) == 0 {
-		return nil, errorsmod.Wrapf(
-			sdkerrors.ErrInvalidRequest, "finality provider BTC public key cannot be empty")
-	}
-
-	fpPK, err := bbn.NewBIP340PubKeyFromHex(req.FpBtcPkHex)
-	if err != nil {
-		return nil, err
-	}
-
-	ctx := sdk.UnwrapSDKContext(c)
-	fp, err := k.GetConsumerFinalityProvider(ctx, req.ConsumerId, fpPK)
-	if err != nil {
-		return nil, err
-	}
-
-	fpResp := types.NewFinalityProviderResponse(fp)
-	return &types.QueryFinalityProviderResponse{FinalityProvider: fpResp}, nil
-}
-
-// FinalityProviderConsumer returns the consumer ID for the finality provider with the specified finality provider BTC PK
-func (k Keeper) FinalityProviderConsumer(c context.Context, req *types.QueryFinalityProviderConsumerRequest) (*types.QueryFinalityProviderConsumerResponse, error) {
-	if req == nil {
-		return nil, status.Error(codes.InvalidArgument, "empty request")
-	}
-
-	if len(req.FpBtcPkHex) == 0 {
-		return nil, errorsmod.Wrapf(
-			sdkerrors.ErrInvalidRequest, "finality provider BTC public key cannot be empty")
-	}
-
-	fpPK, err := bbn.NewBIP340PubKeyFromHex(req.FpBtcPkHex)
-	if err != nil {
-		return nil, err
-	}
-
-	ctx := sdk.UnwrapSDKContext(c)
-	consumerID, err := k.GetConsumerOfFinalityProvider(ctx, fpPK)
-	if err != nil {
-		return nil, err
-	}
-
-	return &types.QueryFinalityProviderConsumerResponse{ConsumerId: consumerID}, nil
 }